refactoring strategies
